Add some explicit casts.
Fixes multiple -Wconversion/-Wsign-compare warnings.

+++ hungrycat.c
@@ -86,7 +86,7 @@ static int eat(const char *filename)
   }
 
 #define check_io(i, j) \
-  while ((i) != (j)) \
+  while ((size_t) (i) != (size_t) (j)) \
   { \
     if ((i) >= 0) \
       errno = EIO; \
@@ -119,7 +119,7 @@ static int eat(const char *filename)
   }
 
   const off_t n_blocks = (file_size + block_size - 1) / block_size;
-  const off_t tail_size = 1 + (file_size - 1) % block_size;
+  const size_t tail_size = (size_t) (1 + (file_size - 1) % block_size);
 
   switch (n_blocks)
   {
@@ -131,7 +131,7 @@ static int eat(const char *filename)
   case 1:
     r_bytes = read(fd, buffer, tail_size);
     check_io(r_bytes, tail_size);
-    w_bytes = write(STDOUT_FILENO, buffer, r_bytes);
+    w_bytes = write(STDOUT_FILENO, buffer, (size_t) r_bytes);
     check_io(w_bytes, r_bytes);
   case 0:
     goto done;
@@ -168,7 +168,7 @@ static int eat(const char *filename)
             break;
           offset += r_bytes;
           fail_if(r_bytes == -1);
-          w_bytes = write(STDOUT_FILENO, buffer, r_bytes);
+          w_bytes = write(STDOUT_FILENO, buffer, (size_t) r_bytes);
           check_io(w_bytes, r_bytes);
           rc = fallocate(fd, FALLOC_FL_KEEP_SIZE | FALLOC_FL_PUNCH_HOLE, 0, offset);
           fail_if(rc != 0);
@@ -194,7 +194,7 @@ static int eat(const char *filename)
     offset = (n_blocks - i - 1) * block_size;
     r_bytes = pread(fd, buffer, block_size, offset);
     check_io(r_bytes, (i == 0 ? tail_size : block_size));
-    w_bytes = pwrite(fd, buffer, r_bytes, i * block_size);
+    w_bytes = pwrite(fd, buffer, (size_t) r_bytes, i * block_size);
     check_io(r_bytes, w_bytes);
     rc = ftruncate(fd, offset);
     fail_if(rc == -1);
@@ -223,7 +223,7 @@ static int eat(const char *filename)
   assert(n_blocks > 0);
   r_bytes = pread(fd, buffer, tail_size, 0);
   check_io(r_bytes, tail_size);
-  w_bytes = write(STDOUT_FILENO, buffer, r_bytes);
+  w_bytes = write(STDOUT_FILENO, buffer, (size_t) r_bytes);
   check_io(w_bytes, r_bytes);
 
 done:
@@ -262,7 +262,7 @@ int main(int argc, char **argv)
           ;
         else if (endptr == optarg || *endptr != '\0')
           errno = EINVAL;
-        else if (value <= 0 || value >= block_size_limit)
+        else if (value <= 0 || (unsigned long) value >= block_size_limit)
           errno = ERANGE;
         if (errno != 0)
         {
@@ -270,7 +270,7 @@ int main(int argc, char **argv)
           show_usage(0);
           return EXIT_FAILURE;
         }
-        block_size = value;
+        block_size = (size_t) value;
         break;
       }
       case 'P':
